Move proto-plakoto logic to a usePlakoto hook
